MealPlan update and stylesheet
