+++ src/screens/MealPlan/styles.js
@@ -1,33 +1,41 @@
-// import { StyleSheet } from "react-native";
 
-// const styles = StyleSheet.create({
-//   container: {
-//     flex: 1,
-//     backgroundColor: "#fff",
-//     padding: 20,
-//   },
-//   title: {
-//     fontSize: 20,
-//     fontWeight: "bold",
-//     marginBottom: 10,
-//   },
-//   recipe: {
-//     flexDirection: "row",
-//     alignItems: "center",
-//     marginBottom: 10,
-//   },
-//   image: {
-//     width: 80,
-//     height: 80,
-//     marginRight: 10,
-//   },
-//   textContainer: {
-//     flex: 1,
-//   },
-//   info: {
-//     fontSize: 16,
-//     marginBottom: 5,
-//   },
-// });
+import {
+    StyleSheet, 
+    Text,
+    View,
+    Image,
+    
+} from 'react-native';
 
-// export default styles;
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      padding: 20,
+    },
+    buttonContainer: {
+      marginBottom: 20,
+    },
+    mealPlanContainer: {
+      marginBottom: 20,
+    },
+    mealTitle: {
+      fontSize: 24,
+      fontWeight: "bold",
+      marginBottom: 10,
+    },
+    image: {
+      height: 200,
+      width: "100%",
+      resizeMode: "cover",
+      marginBottom: 10,
+    },
+    servings: {
+      fontSize: 16,
+      marginBottom: 5,
+    },
+    prepTime: {
+      fontSize: 16,
+    },
+  });
+  
+export default styles;
